Fix: Single file for vizualising, deletion of data_fetcher
